split TesterThreadList to file.
