+++ src/main/scala/chiseltest/internal/BackendInterface.scala
@@ -7,26 +7,6 @@ import chisel3._
 
 import scala.collection.mutable
 
-trait AbstractTesterThread
-
-class TesterThreadList(protected val elts: Seq[AbstractTesterThread]) {
-  def toSeq(): Seq[AbstractTesterThread] = elts
-
-  def join() {
-    Context().backend.doJoin(elts, None)
-  }
-
-  def joinAndStep(clock: Clock) {
-    Context().backend.doJoin(elts, Some(clock))
-  }
-
-  def ++(others: TesterThreadList): TesterThreadList = {
-    new TesterThreadList(elts ++ others.elts)
-  }
-
-  val fork: ForkBuilder = new ForkBuilder(None, None, elts)
-}
-
 /** Common interface definition for tester backends. Internal API.
   */
 trait BackendInterface {
